requirements.txt
discord~=2.3.2
requests~=2.32.5
python-dotenv~=1.1.1
